backend/requirements.txt
pydantic==2.10.2
fastapi==0.110.0
uvicorn==0.30.1
httpx==0.27.0
requests==2.32.3
transformers==4.40.2
accelerate==0.30.1
torch>=2.1.0
huggingface_hub>=0.19.3
numpy>=2.1.0
python-dotenv==1.0.1

